neurartist/requirements.txt
pytorch>=1.1
torchvision
pillow
matplotlib
